+++ config/config.exs
@@ -90,25 +90,30 @@ config :ex_audit,
   ecto_repos: [WraftDoc.Repo],
   version_schema: WraftDoc.ExAudit.Version,
   tracked_schemas: [
+    WraftDoc.BlockTemplates.BlockTemplate,
     WraftDoc.Document.Asset,
-    WraftDoc.Document.BlockTemplate,
     WraftDoc.Document.Block,
-    WraftDoc.Document.Field,
-    WraftDoc.Forms.Form,
     WraftDoc.Document.ContentType,
     WraftDoc.Document.DataTemplate,
-    WraftDoc.Document.Instance.Version,
+    WraftDoc.Document.Field,
     WraftDoc.Document.Instance,
+    WraftDoc.Document.Instance.Version,
     WraftDoc.Document.OrganisationField,
-    WraftDoc.Document.Pipeline.Stage,
     WraftDoc.Document.Pipeline,
-    WraftDoc.Layouts.LayoutAsset,
-    WraftDoc.Layouts.Layout,
+    WraftDoc.Document.Pipeline.Stage,
+    WraftDoc.Document.Theme,
+    WraftDoc.Enterprise.ApprovalSystem,
     WraftDoc.Enterprise.ApprovalSystem,
     WraftDoc.Enterprise.Flow,
+    WraftDoc.Enterprise.Flow,
     WraftDoc.Enterprise.Flow.State,
-    WraftDoc.Themes.Theme,
-    WraftDoc.Enterprise.Vendor
+    WraftDoc.Enterprise.Flow.State,
+    WraftDoc.Enterprise.Vendor,
+    WraftDoc.Enterprise.Vendor,
+    WraftDoc.Forms.Form,
+    WraftDoc.Layouts.Layout,
+    WraftDoc.Layouts.LayoutAsset,
+    WraftDoc.Themes.Theme
   ],
   primitive_structs: [
     Date
