+++ lib/wraft_doc/user/user.ex
@@ -51,7 +51,7 @@ defmodule WraftDoc.Account.User do
     has_many(:user_roles, WraftDoc.Account.UserRole)
     has_many(:roles, through: [:user_roles, :role])
 
-    has_many(:block_templates, WraftDoc.Document.BlockTemplate, foreign_key: :creator_id)
+    has_many(:block_templates, WraftDoc.BlockTemplates.BlockTemplate, foreign_key: :creator_id)
     has_many(:comments, WraftDoc.Comments.Comment)
     has_many(:approval_systems, WraftDoc.Enterprise.ApprovalSystem, foreign_key: :creator_id)
